some minor refactoring for the test environment.
This is a performance micro-optimization, I just happened to notice 
while working on something else. By compiling this as a method that 
returns a literal depending on the presence of the environment variable 
(which is ok because specs are recompiled every time anyway), `if 
testing?` will compile to literally `if true` or `if false`.