Major revision; BREAKING CHANGES
This commit adds semi-tested support for subtitles. The reason this is a 
major, breaking change, is that it changes how files are accessed within 
directories, and how subfiles are stored and organized in memory. I 
think it's definitely for the better, but there are changes all 
throughout the codebase in this commit.